docker/docker-requirements.txt
python-dotenv
prefect-gcp
pandas-gbq
pandas
protobuf
pyarrow
dbt-core
dbt-bigquery